# port/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tableplan API
  version: 1.0.0-spring
  description: Checked-in contract for the Spring Boot Tableplan runtime.
servers:
  - url: /
security:
  - sessionCookie: []
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Tableplan API key with endpoint-specific scopes.
    sessionCookie:
      type: apiKey
      in: cookie
      name: TABLEPLAN_SESSION
    publicShareCookie:
      type: apiKey
      in: cookie
      name: TABLEPLAN_SHARE
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, maxLength: 200 }
    RecipeId:
      name: recipeId
      in: path
      required: true
      schema: { type: string, maxLength: 200 }
...
  schemas:
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string, format: uuid }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
...
    RecipeSummary:
      type: object
      required: [id, sourceId, name, description, tags, ingredients, qualityFlags, visibility, origin, isOwner]
      properties:
        id: { type: string }
        sourceId: { type: string }
        name: { type: string }
        description: { type: string }
        servings: { type: [number, "null"] }
        tags: { type: array, items: { type: string } }
        ingredients: { type: array, items: { type: string } }
        qualityFlags: { type: array, items: { type: string } }
        visibility: { enum: [catalog, user_private, household] }
        origin: { enum: [dataset, manual, paste, upload] }
        isOwner: { type: boolean }
    RecipeSearchResult:
      type: object
      required: [recipes, hasMore, total, limit, offset, nextCursor]
      properties:
        recipes: { type: array, items: { $ref: "#/components/schemas/RecipeSummary" } }
        hasMore: { type: boolean }
        total:
          oneOf:
            - type: object
              required: [value, relation]
              properties:
                value: { type: integer, minimum: 0 }
                relation: { enum: [exact, lowerBound] }
            - type: "null"
        limit: { type: integer }
        offset: { type: integer }
        nextCursor: { type: [string, "null"] }
    MealPlan:
      type: object
      required: [id, name, startsOn, endsOn, version, items]
      properties:
        id: { type: string }
        name: { type: string }
        startsOn: { type: string, format: date }
        endsOn: { type: string, format: date }
        version: { type: integer, format: int64 }
        items: { type: array, items: { type: object } }
...
  responses:
    BadRequest:
      description: Invalid request.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Forbidden:
      description: Principal, role, scope, CSRF, or Origin is denied.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource is absent or not visible.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
paths:
...
  /api/v1/recipes/search:
    get:
      tags: [recipes]
      operationId: searchRecipes
      security: []
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 500 } }
        - { name: ingredient, in: query, schema: { type: string, maxLength: 200 } }
        - { name: tag, in: query, style: form, explode: true, schema: { type: array, maxItems: 12, items: { type: string } } }
        - { name: tagMatch, in: query, schema: { enum: [all, any] } }
        - { name: scope, in: query, schema: { enum: [all, catalog, mine, household] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, maximum: 100000 } }
        - { name: cursor, in: query, schema: { type: string, maxLength: 2000 } }
      responses:
        "200": { description: Recipe results., content: { application/json: { schema: { $ref: "#/components/schemas/RecipeSearchResult" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/v1/recipes/{recipeId}:
    parameters: [{ $ref: "#/components/parameters/RecipeId" }]
    get:
      tags: [recipes]
      operationId: getRecipe
      security: []
      responses: { "200": { description: Recipe detail. }, "404": { $ref: "#/components/responses/NotFound" } }
    put:
      tags: [recipes]
      operationId: updateOwnedRecipe
      responses: { "200": { description: Recipe updated. }, "404": { $ref: "#/components/responses/NotFound" } }
  /api/v1/recipes/{id}/pdf:
    get:
      tags: [recipes, delivery]
      operationId: exportRecipePdf
      parameters: [{ $ref: "#/components/parameters/Id" }, { name: paper, in: query, schema: { enum: [a4, letter], default: a4 } }]
      responses: { "200": { description: PDF document., content: { application/pdf: { schema: { type: string, contentEncoding: binary } } } } }
...
  /api/v1/household/invitations:
    get:
      tags: [household]
      operationId: listHouseholdInvitations
      responses: { "200": { description: Pending invitations for the active household. }, "403": { $ref: "#/components/responses/Forbidden" } }
    post:
      tags: [household]
      operationId: createHouseholdInvitation
      responses: { "200": { description: Invitation; raw token appears once. }, "403": { $ref: "#/components/responses/Forbidden" } }
  /api/v1/household/invitations/{id}:
    delete:
      tags: [household]
      operationId: revokeHouseholdInvitation
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { description: Invitation revoked. } }
  /api/public/household-invitations/{token}:
    get:
      tags: [household]
      operationId: inspectHouseholdInvitation
      security: []
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Safe invitation metadata. }, "404": { $ref: "#/components/responses/NotFound" } }
  /api/v1/household-invitations/{token}/accept:
    post:
      tags: [household]
      operationId: acceptHouseholdInvitation
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Invitation accepted. }, "403": { $ref: "#/components/responses/Forbidden" }, "409": { description: Invitation was already used. } }
  /api/v1/api-keys:
    get:
      tags: [auth]
      operationId: listApiKeys
      responses: { "200": { description: Redacted API keys. } }
    post:
      tags: [auth]
      operationId: createApiKey
      responses: { "200": { description: API key; raw key appears once. }, "400": { $ref: "#/components/responses/BadRequest" } }
  /api/v1/api-keys/{id}:
    delete:
      tags: [auth]
      operationId: revokeApiKey
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { description: API key revoked. } }
  /api/v1/meal-plans:
    get:
      tags: [planning]
      operationId: getMealPlan
      parameters: [{ name: week, in: query, required: true, schema: { type: string, format: date } }]
      responses: { "200": { description: Meal plan or null., content: { application/json: { schema: { oneOf: [{ $ref: "#/components/schemas/MealPlan" }, { type: "null" }] } } } } }
    post:
      tags: [planning]
      operationId: addRecipeToPlan
      responses: { "201": { description: Meal-plan item added. }, "409": { description: Plan changed. } }
  /api/v1/meal-plan-items/{itemId}:
    parameters: [{ name: itemId, in: path, required: true, schema: { type: string } }]
    get:
      tags: [planning]
      operationId: getMealPlanItemContext
      responses: { "200": { description: Meal-plan item and containing plan context. }, "404": { $ref: "#/components/responses/NotFound" } }
    delete:
      tags: [planning]
      operationId: removeMealPlanItem
      responses: { "200": { description: Removal result. } }
    patch:
      tags: [planning]
      operationId: updateMealPlanServings
      responses: { "200": { description: Updated plan. }, "409": { description: Plan changed. } }
  /api/v1/meal-plans/clone-previous:
    post:
      tags: [planning]
      operationId: clonePreviousMealPlan
      responses: { "201": { description: Previous week copied. }, "409": { description: Target is not empty. } }
  /api/v1/meal-plans/{id}/pdf:
    get:
      tags: [planning, delivery]
      operationId: exportMealPlanPdf
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { description: PDF document., content: { application/pdf: { schema: { type: string, contentEncoding: binary } } } } }
  /api/v1/meal-plans/{id}/combined.pdf:
    get:
      tags: [planning, delivery]
      operationId: exportCombinedPdf
      parameters: [{ $ref: "#/components/parameters/Id" }, { name: shoppingListId, in: query, required: true, schema: { type: string } }]
      responses: { "200": { description: Combined plan and shopping PDF. } }
...
  /api/v1/recipe-ingestions:
    post:
      tags: [ingestion]
      operationId: createRecipeIngestion
      requestBody:
        required: true
        content:
          application/json: { schema: { type: object, required: [text], properties: { text: { type: string, maxLength: 102400 }, filename: { type: [string, "null"] } } } }
          multipart/form-data: { schema: { type: object, required: [file], properties: { file: { type: string, contentEncoding: binary } } } }
      responses: { "202": { description: Ingestion queued. }, "413": { description: Artifact rejected. }, "415": { description: Media type is unsupported. } }
  /api/v1/recipe-ingestions/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [ingestion]
      operationId: getRecipeIngestion
      responses: { "200": { description: Safe ingestion state. }, "404": { $ref: "#/components/responses/NotFound" } }
    post:
      tags: [ingestion]
      operationId: publishRecipeIngestion
      responses: { "201": { description: Recipe published. }, "409": { description: Draft is not ready. } }
...